Optionsobject refacctor so that it can save itself now. Read is still done from OptionsModel. XMLdatasources are now split into individual files and folders, and parsed using a static XMLParser class
