+++ Amatsukaze/Application/DatsourceObjects/XMLParsers.cs
@@ -1,47 +1,20 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Diagnostics;
-using System.Reflection;
-using System.ComponentModel;
-using Amatsukaze.HelperClasses;
-using System.Xml;
 using System.IO;
+using System.Linq;
+using System.Reflection;
+using System.Text;
 using System.Text.RegularExpressions;
+using System.Threading.Tasks;
+using System.Xml;
 
 namespace Amatsukaze.ViewModel
 {
-    public interface XMLDataSource
+    public static class XMLParsers
     {
         //Method to parse XML file into an datasourceobject. Throws an exception if some read operation fails.
         //Right now it only reads the first entry... Some name matching might be needed.
-        bool ParseXML(string xmlinput);
-        
-        //Dumps the object to the console
-        void ContentsDump();
-    }
-
-    public class MALDataSource : XMLDataSource
-    {
-        //Properties from MyAnimeList
-        public int id { get; set; }
-        public string title { get; set; }
-        public string english { get; set; }
-        public string synonyms { get; set; }
-        public int episodes { get; set; }
-        public double score { get; set; }
-        public string type { get; set; }
-        public string status { get; set; }
-        public string start_date { get; set; }
-        public string end_date { get; set; }
-        public string synopsis { get; set; }
-        public string image { get; set; }
-
-        //Method to parse XML file into an datasourceobject. Throws an exception if some read operation fails.
-        //Right now it only reads the first entry... Some name matching might be needed.
-        public bool ParseXML(string xmlinput)
+        public static bool MALParseXML(string xmlinput, MALDataSource _MALDataSource)
         {
             try
             {
@@ -93,7 +66,7 @@ namespace Amatsukaze.ViewModel
                         reader.Read();
 
                         //Note: Because of the reflection used, every single property in the animeentry object must have the same name as the property in the xml 
-                        PropertyInfo property = this.GetType().GetProperty(reader.Name);
+                        PropertyInfo property = _MALDataSource.GetType().GetProperty(reader.Name);
                         if (property != null)
                         {
                             if (property.PropertyType == typeof(string))
@@ -101,20 +74,20 @@ namespace Amatsukaze.ViewModel
                                 //only synopsis needs ReadInnerXML because it contains illegal HTML stuff thanks to MAL
                                 if (property.Name == "synopsis")
                                 {
-                                    property.SetValue(this, reader.ReadInnerXml(), null);
+                                    property.SetValue(_MALDataSource, reader.ReadInnerXml(), null);
                                 }
                                 else
                                 {
-                                    property.SetValue(this, reader.ReadElementContentAsString(), null);
+                                    property.SetValue(_MALDataSource, reader.ReadElementContentAsString(), null);
                                 }
                             }
                             else if (property.PropertyType == typeof(int))
                             {
-                                property.SetValue(this, reader.ReadElementContentAsInt(), null);
+                                property.SetValue(_MALDataSource, reader.ReadElementContentAsInt(), null);
                             }
                             else if (property.PropertyType == typeof(double))
                             {
-                                property.SetValue(this, reader.ReadElementContentAsDouble(), null);
+                                property.SetValue(_MALDataSource, reader.ReadElementContentAsDouble(), null);
                             }
                         }
                     }
@@ -129,106 +102,5 @@ namespace Amatsukaze.ViewModel
                 return false;
             }
         }
-        
-        public void ContentsDump()
-        {
-            Type type = this.GetType();
-            PropertyInfo[] properties = type.GetProperties();
-
-            foreach (PropertyInfo property in properties)
-            {
-                Console.WriteLine("{0}: {1}", property.Name, property.GetValue(this, null));
-            }
-        }
-    }
-
-
-    public class AnimeEntryObject : ObservableObjectClass
-    {
-        //Object to hold the properties for one anime (based on the return from myanimelist
-        #region Properties
-
-        //Properties used by Amatsukaze
-        public int id { get; set; }
-        public string title { get; set; }
-        public string english { get; set; }
-        public string synonyms { get; set; }
-        public int episodes { get; set; }
-        public double score { get;set; }
-        public string type { get; set; }
-        public string status { get; set; }
-        public string start_date { get; set; }
-        public string end_date { get; set; }
-        public string synopsis { get; set; }
-        public string image { get; set; }
-
-        //Properties for Amatsukaze GUI
-        public string ImagePath { get; set; }
-
-        private int gridcolumn;
-        public int GridColumn
-        {
-            get
-            {
-                return gridcolumn;
-            }
-            set
-            {
-                if (gridcolumn != value)
-                {
-                    gridcolumn = value;
-                    OnPropertyChanged("GridColumn");
-                }
-            }
-        }      
-
-        private int gridrow { get; set; }
-        public int GridRow
-        {
-            get
-            {
-                return gridrow;
-            }
-            set
-            {
-                if (gridrow != value)
-                {
-                    gridrow = value;
-                    OnPropertyChanged("GridRow");
-                }
-            }
-        }
-
-        //Constructors
-
-        public AnimeEntryObject(MALDataSource datasource)
-        {
-            if (datasource == null)
-            {
-                return;
-            }
-
-            PropertyInfo[] properties = datasource.GetType().GetProperties();
-
-            foreach (PropertyInfo property in properties)
-            {
-                PropertyInfo targetproperty = this.GetType().GetProperty(property.Name);
-                targetproperty.SetValue(this, property.GetValue(datasource, null), null);
-            }
-        }
-
-
-        [Conditional("DEBUG")]
-        public void ContentsDump()
-        {
-            Type type = this.GetType();
-            PropertyInfo[] properties = type.GetProperties();
-
-            foreach (PropertyInfo property in properties)
-            {
-                Console.WriteLine("{0}: {1}", property.Name, property.GetValue(this, null));
-            }
-        }
-        #endregion
     }
 }
