+++ Amatsukaze/Application/OptionsObject.cs
@@ -5,6 +5,9 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Diagnostics;
 using System.IO;
+using Newtonsoft.Json;
+using System.Windows;
+using System.Reflection;
 
 namespace Amatsukaze.ViewModel
 {
@@ -20,6 +23,38 @@ namespace Amatsukaze.ViewModel
         {
             this.Themesetting = "Amatsukaze";
             this.CacheFolderpath = Path.GetDirectoryName(Process.GetCurrentProcess().MainModule.FileName) + @"\Cache\";
+        }
+
+        //The Model class primarily deals with the IO related to saving the user preferences to a JSON file on disk
+        public void Save()
+        {
+            string json = JsonConvert.SerializeObject(this, Formatting.Indented);
+            try
+            {
+                string folderpath = Path.GetDirectoryName(Process.GetCurrentProcess().MainModule.FileName) + @"\Preferences\";
+                string filepath = folderpath + @"\UserPreferences.json";
+
+                FileInfo folder = new FileInfo(folderpath);
+                folder.Directory.Create();
+                File.WriteAllText(filepath, json);
+            }
+            catch (Exception exception)
+            {
+                //FileIO exception
+                MessageBox.Show(exception.Message);
+            }
+        }      
+
+        [Conditional("DEBUG")]
+        public void OptionsDump()
+        {
+            Type type = this.GetType();
+            PropertyInfo[] properties = type.GetProperties();
+
+            foreach (PropertyInfo property in properties)
+            {
+                Console.WriteLine("{0}; {1}", property.Name, property.GetValue(this, null));
+            }
         }
     }
 }
